+++ .github/actions/run-semgrep/.npmrc
@@ -1,0 +1,4 @@
+ignore-scripts=true
+save-exact=true
+audit=true
+fund=false
