+++ .github/workflows/internal_on_push_ci.yml
@@ -4,6 +4,7 @@ on:
   push:
     branches-ignore:
       - main
+  workflow_dispatch:
 
 permissions:
   contents: read
